server/requirements.txt
grpcio
grpcio-tools
protobuf
numpy
asyncio
nest-asyncio
scipy